requirements.txt
# --------- pytorch --------- #
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.12.0+cpu
torchvision==0.13.0+cpu
pytorch-lightning==1.7.1
torchmetrics==0.10.0
timm==0.6.7

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0



# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)


# demo
gradio==3.0
pytorch-lightning-bolts==0.3.2.post1